Refactor MarkupElementViewModel
This refactor makes the MarkupElementViewModel less coupled with the
current HTML conversion logic. As such, the HTML conversion logic will
be more substitutable with another method if need be.
