+++ Dynavity/Dynavity/view-model/HtmlExporter.swift
@@ -1,0 +1,5 @@
+import Combine
+
+protocol HtmlExporter {
+    func getHtmlPublisher(markupText: String, markupType: MarkupElement.MarkupType) -> AnyPublisher<String, Never>
+}
