--- Dynavity/Dynavity/view-model/HtmlRenderable.swift
@@ -1,3 +1,0 @@
-protocol HtmlRenderable {
-    var rawHtml: String { get }
-}
